ライブラリーを更新した＆バージョンナンバーを増やした

+++ Models/WZTextFoldingGenerator.cs
@@ -12,18 +12,18 @@ namespace FooEditor.WinUI.Models
     {
         struct TextLevelInfo
         {
-            public int Index;
+            public long Index;
             public int Level;
-            public TextLevelInfo(int index, int level)
+            public TextLevelInfo(long index, int level)
             {
                 this.Index = index;
                 this.Level = level;
             }
         }
-        public IEnumerable<FoldingItem> AnalyzeDocument(Document doc, int start, int end)
+        public IEnumerable<FoldingItem> AnalyzeDocument(Document doc, long start, long end)
         {
             Stack<TextLevelInfo> beginIndexs = new Stack<TextLevelInfo>();
-            int lineHeadIndex = start;
+            long lineHeadIndex = start;
             foreach (string lineStr in doc.GetLines(start, end))
             {
                 int level = OutlineAnalyzer.GetWZTextLevel(lineStr);
@@ -39,7 +39,7 @@ namespace FooEditor.WinUI.Models
                 yield return item;
         }
 
-        IEnumerable<FoldingItem> GetFoldings(Stack<TextLevelInfo> beginIndexs,int level,int lineHeadIndex)
+        IEnumerable<FoldingItem> GetFoldings(Stack<TextLevelInfo> beginIndexs,int level, long lineHeadIndex)
         {
             while (beginIndexs.Count > 0)
             {
@@ -47,7 +47,7 @@ namespace FooEditor.WinUI.Models
                 if (level > begin.Level)
                     break;
                 beginIndexs.Pop();
-                int endIndex = lineHeadIndex - 1;
+                long endIndex = lineHeadIndex - 1;
                 if (begin.Index < endIndex)
                     yield return new OutlineItem(begin.Index, endIndex,begin.Level);
             }
